app/requirements.txt
avro==1.11.0
confluent-kafka==1.8.2
fastapi==0.75.0
python-dotenv==0.19.2
requests==2.27.1
uvicorn==0.17.6
